+++ Changelog.md
@@ -2,7 +2,7 @@
 This file exists because sometimes I wait to commit my changes until I've done a bulk of work, but I still want to log exactly what's changing. These logs are ordered by date.
 
 ## 2023-03-19
-- Re-set axios request timeout to 90 seconds
+- Re-set axios request timeout to 120 seconds
 - Add invite command
 - Add `/users list` subcommand
 - Add `/invite` command
@@ -15,6 +15,7 @@ This file exists because sometimes I wait to commit my changes until I've done a
 - Create log files
 - Add config schema
 - Add block button to usage requests
+- Improve error handling
 
 ## 2023-03-18
 - Add timestamps and colour to logs